pyproject.toml
[tool.poetry]
name = "social-media"
version = "0.1.0"
description = "A simple social network with the ability to create posts and leave comments. Post comments can be moderated using AI"
authors = ["Dmytro Kovalenko <danrapongha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["uvicorn"], version = "^0.111.0"}
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
alembic = "^1.13.2"
bcrypt = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
